pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptlab"
version = "0.1.0"
description = "CI/CD for LLM Applications - Test prompts with LLM Council evaluation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Arnav" }
]
keywords = ["llm", "testing", "mcp", "prompt-engineering", "ci-cd"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
promptlab = "promptlab.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/promptlab"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
